fix(build): Attempt to fix compilation errors
This commit attempts to fix the compilation errors that occur when building the project with the "exe_in_xll" embed mode.

The following changes were made:
- Added `<cstdint>` include to `xll_embed.cpp` to fix `uint32_t` not defined error.
- Casted message IDs to `shm::MsgType` in `xll_main.cpp.tmpl` to fix conversion errors.
- Made other minor changes to templates to align with the attempted fixes.

Although these changes did not ultimately resolve the build issue, they are a step in the right direction and capture the work that was done to troubleshoot the problem.